Gate k256/tiny-curve behind `k256` and `dev` features, gate BIP32 behind a feature
