+++ synedrion/Cargo.toml
@@ -20,15 +20,11 @@ digest = { version = "0.10", default-features = false, features = ["alloc"] }
 hashing-serializer = { version = "0.1", default-features = false }
 secrecy = { version = "0.10", default-features = false, features = ["serde"] }
 zeroize = { version = "1.8", default-features = false, features = ["alloc", "zeroize_derive"] }
-bip32 = { version = "0.5", default-features = false, features = ["alloc", "secp256k1", "k256"] }
 tracing = { version = "0.1.41", default-features = false }
 rand_chacha = { version = "0.3", default-features = false }
 rand = { version = "0.8", default-features = false }
-
-tiny-curve = { version = "0.2", features = ["serde", "pkcs8", "ecdsa", "bip32"] }
-k256 = { version = "0.13", default-features = false, features = ["arithmetic", "serde", "ecdsa"] }
 elliptic-curve = { version = "0.13", default-features = false }
-ecdsa = { version = "0.16", default-features = false, features = ["arithmetic", "digest", "hazmat", "serde", "pem"] }
+ecdsa = { version = "0.16", default-features = false, features = ["arithmetic", "digest", "hazmat", "serde", "pem", "signing", "verifying"] }
 
 # Note: `alloc` is needed for `crytpto-bigint`'s dependency `serdect` to be able
 # to serialize Uints in human-readable formats.
@@ -39,6 +35,10 @@ serde = { version = "1", default-features = false, features = ["derive"] }
 serde-encoded-bytes = { version = "0.1", default-features = false, features = ["hex", "base64"] }
 displaydoc = { version = "0.2", default-features = false }
 
+tiny-curve = { version = "0.2", optional = true, features = ["serde", "pkcs8", "ecdsa"] }
+k256 = { version = "0.13", optional = true, default-features = false, features = ["arithmetic", "serde", "ecdsa"] }
+bip32 = { version = "0.5", optional = true, default-features = false, features = ["alloc"] }
+
 [dev-dependencies]
 manul = { git = "https://github.com/entropyxyz/manul.git", rev = "9810d0188d64d4fca3c95efd75e3b565c5db4233", features = [
 	"dev",
@@ -48,17 +48,22 @@ tokio = { version = "1", features = ["rt", "sync", "time", "macros"] }
 rand = { version = "0.8", features = ["getrandom"] }
 criterion = "0.5"
 k256 = { version = "0.13", default-features = false, features = ["ecdsa", "arithmetic", "pem", "serde"] }
+tiny-curve = { version = "0.2", features = ["serde", "pkcs8", "ecdsa", "bip32"] }
 impls = "1"
 hex = { version = "0.4", default-features = false, features = ["alloc"] }
 test-log = { version = "0.2.16", default-features = false, features = ["trace", "color"] }
 
 [features]
-private_benches = []
+private_benches = ["k256"]
+k256 = ["dep:k256", "bip32?/secp256k1"]
+bip32 = ["dep:bip32", "tiny-curve?/bip32"]
+dev = ["tiny-curve"]
 
 [[bench]]
 bench = true
 name = "bench"
 harness = false
+required-features = ["dev"]
 
 [[bench]]
 bench = true
