+++ synedrion/src/cggmp21/interactive_signing.rs
@@ -2621,7 +2621,8 @@ mod tests {
 
     use super::InteractiveSigning;
     use crate::{
-        cggmp21::{AuxInfo, KeyShare, TestParams},
+        cggmp21::{AuxInfo, KeyShare},
+        dev::TestParams,
         SchemeParams,
     };
     type Curve = <TestParams as SchemeParams>::Curve;
