+++ synedrion/src/cggmp21/interactive_signing_tests.rs
@@ -19,12 +19,12 @@ use super::{
         Round2EchoBroadcast, Round2NormalBroadcast, Round3, Round3EchoBroadcast, Round3NormalBroadcast, Round3Payload,
         Round4NormalBroadcast, Round5, Round6,
     },
-    params::TestParams,
     sigma::{ElogProof, ElogPublicInputs, ElogSecretInputs},
     SchemeParams,
 };
 use crate::{
     curve::{Point, RecoverableSignature, Scalar},
+    dev::TestParams,
     tools::{
         protocol_shortcuts::{DowncastMap, MapValues},
         protocol_shortcuts_dev::{check_evidence_with_behavior, check_invalid_message_evidence, CheckPart},
