+++ synedrion/src/cggmp21/key_init_tests.rs
@@ -10,11 +10,11 @@ use rand_core::{CryptoRngCore, OsRng};
 
 use super::{
     key_init::{KeyInit, KeyInitAssociatedData, Round2NormalBroadcast, Round3, Round3NormalBroadcast},
-    params::TestParams,
     sigma::SchProof,
 };
 use crate::{
     curve::Scalar,
+    dev::TestParams,
     tools::{
         bitvec::BitVec,
         protocol_shortcuts_dev::{check_evidence_with_behavior, check_invalid_message_evidence, CheckPart},
