+++ synedrion/src/cggmp21/params.rs
@@ -4,30 +4,24 @@ use core::{fmt::Debug, ops::Add};
 // and `k256` depends on the released one.
 // So as long as that is the case, `k256` `Uint` is separate
 // from the one used throughout the crate.
-use crypto_bigint::{NonZero, Uint, U1024, U128, U2048, U256, U4096, U512, U8192};
+use crypto_bigint::NonZero;
 use digest::generic_array::ArrayLength;
 use ecdsa::hazmat::{DigestPrimitive, SignPrimitive, VerifyPrimitive};
 use elliptic_curve::{
-    bigint::{self as bigintv05, Concat, Split, Uint as CurveUint},
+    bigint::{self as bigintv05, Concat, Split},
     point::DecompressPoint,
     sec1::{FromEncodedPoint, ModulusSize, ToEncodedPoint},
     Curve, CurveArithmetic, PrimeCurve, PrimeField,
 };
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 
-use tiny_curve::TinyCurve32;
+use crate::{paillier::PaillierParams, tools::hashing::HashableType};
 
-use crate::{
-    paillier::PaillierParams,
-    tools::hashing::HashableType,
-    uint::{U1024Mod, U128Mod, U2048Mod, U256Mod, U4096Mod, U512Mod},
-};
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PaillierTest;
-
+#[cfg(any(test, feature = "k256", feature = "dev"))]
 #[allow(clippy::indexing_slicing)]
-const fn upcast_uint<const N1: usize, const N2: usize>(value: CurveUint<N1>) -> CurveUint<N2> {
+pub(crate) const fn upcast_uint<const N1: usize, const N2: usize>(
+    value: elliptic_curve::bigint::Uint<N1>,
+) -> elliptic_curve::bigint::Uint<N2> {
     assert!(N2 >= N1, "Upcast target must be bigger than the upcast candidate");
     let mut result_words = [0; N2];
     let mut i = 0;
@@ -36,38 +30,12 @@ const fn upcast_uint<const N1: usize, const N2: usize>(value: CurveUint<N1>) -> 
         result_words[i] = words[i];
         i += 1;
     }
-    CurveUint::from_words(result_words)
+    elliptic_curve::bigint::Uint::from_words(result_words)
 }
 
-const fn convert_uint<const N: usize>(value: CurveUint<N>) -> Uint<N> {
-    Uint::from_words(value.to_words())
-}
-
-impl PaillierParams for PaillierTest {
-    const PRIME_BITS: u32 = 128;
-    type HalfUint = U128;
-    type HalfUintMod = U128Mod;
-    type Uint = U256;
-    type UintMod = U256Mod;
-    type WideUint = U512;
-    type WideUintMod = U512Mod;
-    type ExtraWideUint = U1024;
-}
-
-/// Paillier parameters corresponding to 112 bits of security.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PaillierProduction112;
-
-// Source of the values: Appendix C.1.
-impl PaillierParams for PaillierProduction112 {
-    const PRIME_BITS: u32 = 1024;
-    type HalfUint = U1024;
-    type HalfUintMod = U1024Mod;
-    type Uint = U2048;
-    type UintMod = U2048Mod;
-    type WideUint = U4096;
-    type WideUintMod = U4096Mod;
-    type ExtraWideUint = U8192;
+#[cfg(any(test, feature = "k256", feature = "dev"))]
+pub(crate) const fn convert_uint<const N: usize>(value: elliptic_curve::bigint::Uint<N>) -> crypto_bigint::Uint<N> {
+    crypto_bigint::Uint::from_words(value.to_words())
 }
 
 /// Signing scheme parameters.
@@ -125,60 +93,11 @@ where
     }
 }
 
-/// Scheme parameters **for testing purposes only**.
-/// Security is weakened to allow for faster execution.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize, PartialOrd, Ord)]
-pub struct TestParams;
-
-impl SchemeParams for TestParams {
-    type Curve = TinyCurve32;
-    // TODO: ReprUint is typenum::U192 because of RustCrypto stack internals, hence the U384 here,
-    // but once that is solved, this can be a U128 (or even smaller).
-    type WideCurveUint = bigintv05::U384;
-    const SECURITY_BITS: usize = 16;
-    const SECURITY_PARAMETER: usize = 32;
-    const L_BOUND: u32 = 32;
-    const EPS_BOUND: u32 = 64;
-    const LP_BOUND: u32 = 160;
-    type Paillier = PaillierTest;
-    const CURVE_ORDER: NonZero<<Self::Paillier as PaillierParams>::Uint> =
-        convert_uint(upcast_uint(Self::Curve::ORDER))
-            .to_nz()
-            .expect("Correct by construction");
-    const CURVE_ORDER_WIDE: NonZero<<Self::Paillier as PaillierParams>::WideUint> =
-        convert_uint(upcast_uint(Self::Curve::ORDER))
-            .to_nz()
-            .expect("Correct by construction");
-}
-
-/// Production strength parameters.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize, Ord, PartialOrd)]
-pub struct ProductionParams112;
-
-impl SchemeParams for ProductionParams112 {
-    type Curve = k256::Secp256k1;
-    type WideCurveUint = bigintv05::U512;
-    const SECURITY_BITS: usize = 112;
-    const SECURITY_PARAMETER: usize = 256;
-    const L_BOUND: u32 = 256;
-    const EPS_BOUND: u32 = Self::L_BOUND * 2;
-    const LP_BOUND: u32 = Self::L_BOUND * 5;
-    type Paillier = PaillierProduction112;
-    const CURVE_ORDER: NonZero<<Self::Paillier as PaillierParams>::Uint> =
-        convert_uint(upcast_uint(Self::Curve::ORDER))
-            .to_nz()
-            .expect("Correct by construction");
-    const CURVE_ORDER_WIDE: NonZero<<Self::Paillier as PaillierParams>::WideUint> =
-        convert_uint(upcast_uint(Self::Curve::ORDER))
-            .to_nz()
-            .expect("Correct by construction");
-}
-
 #[cfg(test)]
 mod tests {
     use super::{
         bigintv05::{U256, U64},
-        upcast_uint, ProductionParams112, SchemeParams, TestParams,
+        upcast_uint,
     };
 
     #[test]
@@ -202,11 +121,5 @@ mod tests {
         let n256 = U256::from_u8(8);
         let n: U256 = upcast_uint(n256);
         assert_eq!(n, n256)
-    }
-
-    #[test]
-    fn parameter_consistency() {
-        assert!(ProductionParams112::are_self_consistent());
-        assert!(TestParams::are_self_consistent());
     }
 }
