+++ synedrion/src/curve.rs
@@ -3,10 +3,17 @@
 //! and we isolate all the related logic here.
 
 mod arithmetic;
-mod bip32;
 mod ecdsa;
+
+#[cfg(feature = "bip32")]
+mod bip32;
 
 pub(crate) use arithmetic::{secret_split, Point, Scalar};
 
 pub use self::ecdsa::RecoverableSignature;
-pub use bip32::DeriveChildKey;
+
+#[cfg(feature = "bip32")]
+pub use bip32::{DeriveChildKey, PublicTweakable, SecretTweakable};
+
+#[cfg(feature = "bip32")]
+pub(crate) use bip32::{apply_tweaks_public, derive_tweaks};
