+++ synedrion/src/lib.rs
@@ -13,6 +13,16 @@
 )]
 #![cfg_attr(not(test), warn(clippy::unwrap_used, clippy::indexing_slicing))]
 
+/*!
+## Features
+
+`k256`: Secp256k1 parameters using [`k256`](`::k256`) crate. See the [`k256`] module.
+
+`dev`: Non-secure development parameters using [`tiny-curve`](`::tiny_curve`) crate. See the [`dev`] module.
+
+`bip32`: enables BIP32 support for [`ThresholdKeyShare`].
+*/
+
 extern crate alloc;
 
 mod cggmp21;
@@ -22,17 +32,31 @@ mod tools;
 mod uint;
 mod www02;
 
+#[cfg(feature = "k256")]
+pub mod k256;
+
+#[cfg(any(test, feature = "dev"))]
+pub mod dev;
+
+#[cfg(test)]
+mod tests;
+
 // Some re-exports to avoid the need for version-matching
+#[cfg(feature = "bip32")]
 pub use bip32;
+
 pub use signature;
 
 pub use cggmp21::{
     AuxGen, AuxGenAssociatedData, AuxGenProtocol, AuxInfo, InteractiveSigning, InteractiveSigningAssociatedData,
     InteractiveSigningProtocol, KeyInit, KeyInitAssociatedData, KeyInitProtocol, KeyRefresh, KeyRefreshAssociatedData,
-    KeyRefreshProtocol, KeyShare, KeyShareChange, PrehashedMessage, ProductionParams112, SchemeParams, TestParams,
+    KeyRefreshProtocol, KeyShare, KeyShareChange, PrehashedMessage, SchemeParams,
 };
-pub use curve::{DeriveChildKey, RecoverableSignature};
+pub use curve::RecoverableSignature;
 pub use www02::{KeyResharing, KeyResharingProtocol, NewHolder, OldHolder, ThresholdKeyShare};
+
+#[cfg(feature = "bip32")]
+pub use curve::{DeriveChildKey, PublicTweakable, SecretTweakable};
 
 #[cfg(feature = "private_benches")]
 #[allow(missing_docs)]
