+++ synedrion/src/private_benches.rs
@@ -10,9 +10,9 @@ use crate::{
             DecProof, DecPublicInputs, DecSecretInputs, ElogProof, ElogPublicInputs, ElogSecretInputs, EncElgProof,
             EncElgPublicInputs, EncElgSecretInputs, FacProof, ModProof, PrmProof, SchCommitment, SchProof, SchSecret,
         },
-        ProductionParams112,
     },
     curve::Scalar,
+    k256::ProductionParams112,
     paillier::{Ciphertext, PaillierParams, RPParams, RPSecret, Randomizer, SecretKeyPaillier, SecretKeyPaillierWire},
     tools::Secret,
     uint::SecretSigned,
