+++ synedrion/src/uint.rs
@@ -6,7 +6,4 @@ mod traits;
 pub(crate) use public_signed::PublicSigned;
 pub(crate) use secret_signed::SecretSigned;
 pub(crate) use secret_unsigned::SecretUnsigned;
-pub(crate) use traits::{
-    Exponentiable, FromXofReader, HasWide, IsInvertible, ToMontgomery, U1024Mod, U128Mod, U2048Mod, U256Mod, U4096Mod,
-    U512Mod,
-};
+pub(crate) use traits::{Exponentiable, FromXofReader, HasWide, IsInvertible, ToMontgomery};
